oauth2/srcs/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
gunicorn==22.0.0
idna==3.7
packaging==24.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.1
